csharp: Fix initialization of inherit Eo class
They should be stored in each wrapped Eo class instead of a single
static field of a root class. This was causing the efl_super calls in
inherited classes to fail, as the actual Eo_Class pointed was the root
one.

Now the target class is passed by ref to the root constructor to be set
to the newly created Eo_Class if it is in an inherited context.